Summary (required)
Made changes throughout the stylesheet in order to have a mobile view below 768px. Tablet view 768px-1023px, and PC view 1024px and above.

Changed project to a mobile first view and added the media quiries in order to acomodate to other devices.

Changed the display to inline-block for the profile images and added them in the media quiries in order to reflect the mock images for mobile, tablet, and pc screen view.